+++ R/plot.emm.R
@@ -35,6 +35,10 @@ plot.emmGrid = function(x, y, type, intervals = TRUE, comparisons = FALSE,
         object = update(x, predict.type = type, ..., silent = TRUE)
     else
         object = update(x, ..., silent = TRUE)
+    # when we want comparisons, we have a transformation, and we want non-link scale, it's mandatory to regrid first:
+    if(comparisons && !is.null(object@misc$tran) && 
+            !(object@misc$predict.type %in% c("link", "lp", "linear.predictor")))
+        object = regrid(object, transform = object@misc$predict.type)
     if (missing(int.adjust)) {
         int.adjust = object@misc$adjust
         if (is.null(int.adjust))